Metal spis (#1)
* Add a configuration parameter to control the use of Metal SPIs.

With the new `MVK_CONFIG_USE_METAL_PRIVATE_API`, the user can prevent their
use at run time without recompiling MoltenVK. This may be useful for
troubleshooting purposes.

The extra member is because the compiler adds tail padding to make the
struct aligned on an 8-byte boundary, so the `static_assert()` fails
without it.

* MVKPipeline: Support framebuffer logic operations.

This support was always there in Metal. It was just hidden from public
view--but not from prying eyes on the Objective-C metadata.

I had initially thought they were using OpenGL/Vulkan values for the
logic operation. When that didn't work, I then tried D3D11 values. I
guess D3D11 makes sense, because the Intel GPU drivers were the first to
support this in Metal.

Move the definition of `MVK_USE_METAL_PRIVATE_API` into
`"MVKCommonEnvironment.h"` so we can use it from non-C++ code.

* MVKSampler: Support LoD bias.

Metal actually supports this, but this is a hidden property of the
`MTLSamplerDescriptor`. It's not part of the public API. Guard our use
of it appropriately.

* MVKPipeline: Support sample masks in the Metal SPI.

This is yet another "secret" property on the
`MTLRenderPipelineDescriptor`.

* Support the depth bounds test.

This isn't even a regular SPI. It's not even present in the Metal
framework. It's exposed by the driver. Only AMD drivers support the
method we need for now.

---------

Co-authored-by: Chip Davis <chip@holochip.com>
Co-authored-by: Chip Davis <cdavis@codeweavers.com>